Fixed bugs

+++ models/model_builder.py
@@ -43,7 +43,9 @@ class ExtSummarizer(nn.Module):
 
     def forward(self, src, segs, clss, mask_src, mask_cls):
         top_vec = self.bert(src, segs, mask_src)
-        sents_vec = top_vec[torch.arange(top_vec.size(0)).unsqueeze(1), clss]
+        # print(torch.arange(top_vec.size(0)).unsqueeze(1).size())
+        sents_vec = top_vec[torch.arange(top_vec.size(0)).unsqueeze(1).long(), clss.long()]
+        # print(sents_vec.size())
         sents_vec = sents_vec * mask_cls[:, :, None].float()
         sent_scores = self.ext_layer(sents_vec, mask_cls).squeeze(-1)
         return sent_scores, mask_cls
